backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
spacy>=3.7
scikit-learn>=1.3
gunicorn>=21.2
whitenoise>=6.6
python-dotenv>=1.0
dj-database-url>=2.1
